+++ NEWS.md
@@ -1,5 +1,8 @@
 # surveytable (development version)
 
+* Optionally, show the test statistic, control rounding of the p-value.
+* Improved handling of missing values and the `drop_na` argument in `tab_subset()` and `tab_subset_rate()`.
+
 # surveytable 0.9.7
 
 * Bugfixes.